API/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
Pillow
svgwrite
fpdf
aiohttp
requests
